Q-TCP Added
Q Learning agent added. The reward function is based on instantaneous delay and throughput of the sender. Input to the learner: the state (i.e., EWMA of the ACK inter-arrival time, EWMA of the packet inter-sending time, the ratio of current RTT and the minimum RTT, the slow start threshold, and cwnd size), and reward r from the network. It is based on the paper [Improving TCP Congestion Control with Machine Intelligence](https://dl.acm.org/doi/10.1145/3229543.3229550).